pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "flask-limiter>=3.8.0",
    "flask-sqlalchemy>=3.1.1",
    "flask>=3.1.0",
    "psycopg2-binary>=2.9.10",
    "user-agents>=2.2.0",
]
